Update index.tsx
[UPDATED as of 3/10/24 3:34 PST]
Corrected grammar issues on the websites, revised the wording of the landing page, and added a website logo featuring a cup of tea as the tab icon. Updated the copyright notice at the bottom to 2024. Additionally, updated front-page text to include more professional language and corrected errors in the review section, where Network Chuck was mistakenly labeled as DBTech, along with many other edits in wording and grammar.