Updated index.tsx
[UPDATED as of 3/10/24 3:34 PST]
I corrected grammar issues on the websites, revised the wording of the landing page, and added a logo featuring a cup of tea as the tab icon. I updated the copyright notice at the bottom to 2024. Additionally, I updated the front-page text to include more professional language. I corrected errors in the review section, where Network Chuck was mistakenly labeled as DBTech, along with many other edits in wording and grammar.